+++ src/collocations_mt_.cpp
@@ -1,5 +1,5 @@
-#include "quanteda.h"
-#include "dev.h"
+#include "../include/quanteda.h"
+#include "../include/dev.h"
 #include <bitset>
 using namespace quanteda;
