+++ src/dist_mt.cpp
@@ -1,5 +1,5 @@
 #include "armadillo.h"
-#include "quanteda.h"
+#include "../include/quanteda.h"
 using namespace quanteda;
 using std::pow;
 using std::exp;
