+++ src/fcm_mt.cpp
@@ -1,5 +1,5 @@
-#include "quanteda.h"
-#include "dev.h"
+#include "../include/quanteda.h"
+#include "../include/dev.h"
 using namespace quanteda;
 
 struct hash_pair {
