+++ src/tokens_chunk_mt.cpp
@@ -1,6 +1,6 @@
-#include "quanteda.h"
-#include "recompile.h"
-//#include "dev.h"
+#include "../include/quanteda.h"
+#include "../include/recompile.h"
+//#include "../include/dev.h"
 using namespace quanteda;
 
 Texts chunk(Text &tokens,
