+++ src/tokens_select_mt.cpp
@@ -1,6 +1,6 @@
-#include "quanteda.h"
-//#include "dev.h"
-#include "recompile.h"
+#include "../include/quanteda.h"
+//#include "../include/dev.h"
+#include "../include/recompile.h"
 using namespace quanteda;
